% Generated by roxygen2 (4.0.1): do not edit by hand
\name{absolute_path}
\alias{absolute_path}
\title{Check is an URL is absolute or relative}
\usage{
absolute_path(url)
}
\arguments{
\item{url}{the url(s) to check}
}
\value{
Returns a logical vector indicating whether the path is absolute or relative.
}
\description{
Check is an URL is absolute or relative
}

