% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dpattr}
\alias{dpattr}
\alias{dpattr.datapackage}
\alias{dpattr.dpfield}
\alias{dpattr.dpresource}
\alias{dpdescription}
\alias{dpformat}
\alias{dpname}
\alias{dptitle}
\alias{dptype}
\title{Get and set the attributes in datapackage meta data}
\usage{
dpattr(meta, attribute, ...)

\method{dpattr}{datapackage}(meta, attribute, ...)

\method{dpattr}{dpresource}(meta, attribute, ...)

\method{dpattr}{dpfield}(meta, attribute, ...)

dptitle(meta, ...)

dpname(meta, ...)

dpdescription(meta, ...)

dptype(meta, ...)

dpformat(meta, ...)
}
\arguments{
\item{meta}{the meta data for which the attribute needs to be set or get.
This can be an object of type \code{\link{datapackage}},
\code{\link{dpresource}} or \code{\link{dpfield}}.}

\item{attribute}{the name of the attribute that needs to be set.}

\item{...}{currently unused.}
}
\description{
Get and set the attributes in datapackage meta data
}
\details{
A \code{datapackage} contains one or more resources (\code{dpresource}),
which contain a number of fields (\code{dpfield}). Each of these objects
contain a number of attributes such as a name, title and description, that
describe the object. These attributes can be read and modified using the
methods presented here.

The objects \code{datapackage}, \code{dpresource} and \code{dpfield} are just
lists and can besides the methods presented here, also easily be modified
using the regular list operation, e.g. to change the title of an
\code{datapackage} one could also use \code{meta[['title']] <- "My title"}.

The methods \code{dptitle}, \code{dpname}, \code{dpdescription}, \code{type}
and \code{format} are just shortcuts for \code{dpattr('title')} etc.
}

