% Generated by roxygen2 (4.0.2): do not edit by hand
\name{datapackage}
\alias{datapackage}
\title{Create a \code{\link{datapackage}} object}
\usage{
datapackage(name, title, description = "", data)
}
\arguments{
\item{name}{the name of the datapackage}

\item{title}{the title of the datapackage}

\item{description}{the description of the datapackage}

\item{data}{a \code{data.frame} that is added as resource to the datapackage.}
}
\value{
A \code{datapackage} object.
}
\description{
Create a \code{\link{datapackage}} object
}

