% Generated by roxygen2 (4.0.2): do not edit by hand
\name{download_file}
\alias{download_file}
\title{Download a file from a URL into a local temporary file}
\usage{
download_file(url)
}
\arguments{
\item{url}{a character vector containing the url to the file. When the
  protocal (\code{http}, \code{ftp}, \code{file} or \code{https}) is missing,
  it is assumed that the url points to a local file; and the protocal
  \code{file} is used.}
}
\value{
Returns the name of the temporary file.
}
\description{
Download a file from a URL into a local temporary file
}
\details{
Uses \code{\link{download.file}} to download the file; except when the URL
points to an \code{https} address. In that case the package \code{RCurl} is
used.
}

