% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dpadd_resource}
\alias{dpadd_resource}
\title{Add a resource to a \code{\link{datapackage}}}
\usage{
dpadd_resource(meta, data, name, title, description = "", replace = FALSE,
  inline_data = FALSE)
}
\arguments{
\item{meta}{a \code{\link{datapackage}} object}

\item{data}{a \code{data.frame} that is added as resource to the datapackage.}

\item{name}{the name of the resource.}

\item{title}{the title of the resource.}

\item{description}{a description of the resource.}

\item{replace}{replace an existing data resource with the same name (if such a
resource exists).}

\item{inline_data}{add the data to the resource, not just the description of
  the data.}
}
\value{
Returns the original \code{meta} object with the new resource added.
}
\description{
Add a resource to a \code{\link{datapackage}}
}

