% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dpdata}
\alias{dpdata}
\title{Retrieve the data of  a resource}
\usage{
dpdata(meta, resource = 1)
}
\arguments{
\item{meta}{a \code{datapackage} object or a \code{\link{dpresource}} object
in which case the \code{resource} argument is ignored.}

\item{resource}{an identifier of the the resource. This can be a numeric
  index, or a character with the name of the resource.}
}
\value{
A \code{data.frame} containing the data of the resource.
}
\description{
Retrieve the data of  a resource
}
\details{
Currently the following types are supported:
\describe{
  \item{string}{a string (of arbitrary length)}
  \item{number}{a number including floating point numbers.}
  \item{integer}{an integer.}
  \item{date}{a date. This MUST be in ISO6801 format YYYY-MM-DD or, if not, a
  format field must be provided describing the structure.}
  \item{datetime}{a date-time. This MUST be in ISO 8601 format of
  YYYY-MM-DDThh:mm:ssZ in UTC time or, if not, a format field must be
  provided.}
  \item{boolean}{a boolean value (1/0, true/false).}
}
}

