% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dpfield}
\alias{dpfield}
\alias{dpfield<-}
\title{Get and set the meta data of a single field}
\usage{
dpfield(meta, field, resource = 1)

dpfield(meta, field, resource = 1) <- value
}
\arguments{
\item{meta}{a \code{\link{datapackage}} object or a \code{\link{dpresource}}
object in which case the \code{resource} argument is ignored.}

\item{field}{the name or number of the field.}

\item{resource}{an identifier of the the resource. This can be a numeric
index, or a character with the name of the resource.}

\item{value}{a \code{dpfield} object to assign to the field.}
}
\value{
\code{dpfield} returns an object of type \code{dpfield}.
}
\description{
Get and set the meta data of a single field
}

