% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dpfields}
\alias{dpfields}
\title{Get list of fields in the file}
\usage{
dpfields(meta, resource = 1)
}
\arguments{
\item{meta}{a \code{\link{datapackage}} object or a \code{\link{dpresource}}
object in which case the \code{resource} argument is ignored.}

\item{resource}{an identifier of the the resource. This can be a numeric
  index, or a character with the name of the resource.}
}
\value{
Returns a character vector with field names.
}
\description{
Get list of fields in the file
}

