% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dpopen}
\alias{dpopen}
\title{Open a data package}
\usage{
dpopen(datapackage)
}
\arguments{
\item{datapackage}{a directory containing a datapackage.}
}
\value{
Return a \code{\link{datapackage}} object.
}
\description{
Open a data package
}

