% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dpresource}
\alias{dpresource}
\alias{dpresource<-}
\title{Get or set the meta data of a specific resource from a datapackage}
\usage{
dpresource(meta, resource = 1)

dpresource(meta, resource = 1) <- value
}
\arguments{
\item{meta}{a \code{\link{datapackage}} object}

\item{resource}{an identifier of the resource. This can be a numeric
index, or a character with the name of the resource.}

\item{value}{a \code{dpresource} object}
}
\value{
\code{dpresource} returns a \code{dpresource} object.
}
\description{
Get or set the meta data of a specific resource from a datapackage
}

