% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dpresources}
\alias{dpresources}
\title{Get list of the resources in the datapackage.}
\usage{
dpresources(meta)
}
\arguments{
\item{meta}{a \code{\link{datapackage}} object}
}
\value{
Returns a character vector with the resource names
}
\description{
Get list of the resources in the datapackage.
}

