% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dpsave}
\alias{dpsave}
\title{Save a datapackage with resources}
\usage{
dpsave(meta, path, ...)
}
\arguments{
\item{meta}{the \code{\link{datapackage}} object to save.}

\item{path}{the directory in which to save the \code{\link{datapackage}}}

\item{...}{the data of the resources. See details.}
}
\description{
Save a datapackage with resources
}
\details{
Since the data belonging to the resources is not stored in the
\code{\link{datapackage}} object. the data of each of the resources in the
\code{\link{datapackage}} needs to be passed to the \code{dpsave} function
as additional argument as \code{<resourcename>=<data.frame>}.

When \code{path} does not exist it is created. Any files existing in
\code{path} are overwritten.
}

