% Generated by roxygen2 (4.0.1): do not edit by hand
\name{format_column}
\alias{format_column}
\title{Convert a data column to character for saving}
\usage{
format_column(x, meta)
}
\arguments{
\item{x}{the column to convert.}

\item{meta}{the \code{\link{dpfield}} object belonging to the field.}
}
\value{
The input column converted to character column.
}
\description{
Convert a data column to character for saving
}

