% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.datapackage}
\alias{print.datapackage}
\title{Print a \code{\link{datapackage}} object}
\usage{
\method{print}{datapackage}(x, ...)
}
\arguments{
\item{x}{a \code{\link{datapackage}} object}

\item{...}{ignored.}
}
\description{
Print a \code{\link{datapackage}} object
}

