% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.dpfield}
\alias{print.dpfield}
\title{Print a \code{\link{dpfield}} object}
\usage{
\method{print}{dpfield}(x, ...)
}
\arguments{
\item{x}{a \code{\link{dpfield}} object}

\item{...}{ignored.}
}
\description{
Print a \code{\link{dpfield}} object
}

