% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.dpresource}
\alias{print.dpresource}
\title{Print a \code{\link{dpresource}} object}
\usage{
\method{print}{dpresource}(x, ...)
}
\arguments{
\item{x}{a \code{\link{dpresource}} object}

\item{...}{ignored.}
}
\description{
Print a \code{\link{dpresource}} object
}

