% Generated by roxygen2 (4.0.1): do not edit by hand
\name{resource_index}
\alias{resource_index}
\title{Get the index of a resource in the \code{\link{datapackage}}}
\usage{
resource_index(meta, resource)
}
\arguments{
\item{meta}{a \code{\link{datapackage}} object}

\item{resource}{an identifier of the the resource. This can be a numeric
index, or a character with the name of the resource.}
}
\description{
Returns a numeric index. In case of a valid numeric value of \code{resource}
\code{resource} is returned. In case of a character value of \code{resource}
the index of the resource with the same name is returned. In case the
resource could not be found the value \code{NULL} is returned.
}

