% Generated by roxygen2 (4.0.2): do not edit by hand
\name{save_resource}
\alias{save_resource}
\title{Save the data of a resource to file}
\usage{
save_resource(meta, resource, data, filename)
}
\arguments{
\item{meta}{a \code{datapackage} object or a \code{\link{dpresource}} object
in which case the \code{resource} argument is ignored.}

\item{resource}{an identifier of the the resource. This can be a numeric
index, or a character with the name of the resource.}

\item{data}{the data of the resource. When omitted the \code{data} element
of the resource is used (if present).}

\item{path}{folder in which to save the data.}
}
\value{
Returns the \code{datapackage} object with the new path, and filename of the
resource.
}
\description{
Save the data of a resource to file
}

