% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dpfield}
\alias{dpfield}
\title{Get the meta data of a single field}
\usage{
dpfield(meta, field, resource = 1)
}
\arguments{
\item{meta}{a \code{\link{datapackage}} object or a \code{\link{dpresource}}
object in which case the \code{resource} argument is ignored.}

\item{resource}{an identifier of the the resource. This can be a numeric
index, or a character with the name of the resource.}

\item{field}{the name or number of the field.}
}
\value{
Returns an object of type \code{dpfieldmeta}.
}
\description{
Get the meta data of a single field
}

