% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dpresource}
\alias{dpresource}
\title{Get the meta data of a specific resource from a datapackage}
\usage{
dpresource(meta, resource = 1)
}
\arguments{
\item{meta}{a \code{\link{datapackage}} object}

\item{resource}{an identifier of the the resource. This can be a numeric
  index, or a character with the name of the resource.}
}
\value{
Returns a \code{dpresource} object.
}
\description{
Get the meta data of a specific resource from a datapackage
}

